fix(m3): PR #32 一轮审计(cubic 17 条采纳 13 申辩 4 + Codex 2 条采纳 2)——P1 页宿主尺寸分量错(PAGE.2/3 是宽高非右下,dlu(PAGE.2-PAGE.0) 把 194 宽算成 88;改 dlu(PAGE.2,PAGE.3) 直传+宿主取模板最大 194x216)+ P2 框架缺 WS_CAPTION|WS_SYSMENU|WS_CLIPCHILDREN(rc STYLE 逐字——无标题栏无 X,冒烟只验了窗口文本属性未验可见性)+ P2 OK/Cancel 按钮坐标分量错(bx2 取了 dlu 的纵向分量,Cancel 横坐标错)+ P2 左键未实现值误入拖拽(上游 per-value switch 各值各事,未知值无事;改 0 落穿/_ return,双击同理 0|1|2|5|6 显式切全屏)+ P2 auto_fit 乘法 i32 溢出(i64 加宽再除)+ P2 exe 位于 %APPDATA%\riviv 时标记写覆盖同路径全表(exe_dir==appdata_dir 跳过标记写)+ P2 fit 变更后视图不重锚定(双 bot 同报;上游 Options 路径确无 on_size 但其 FILL WINDOW 菜单命令成对调 _viv_on_size viv.c:2032-2051——Effects 增 refit 旗标,OK 时 set_view 重锚定后再 invalidate)+ P2 1:1 离开 pos 不变不重绘(fill 上采样下 level0 尺寸已变;was_one_to_one 计为变动)+ P2 frame_minus 翻转状态栏不即时(OK 无条件 refresh_status;SB_SETTEXT 缓存去重)+ P3 menu.rs ViewOptions 过期注释/fit_size 悬空行/&F 助记符冲突(Fill Window 复用菜单串去助记符,保 Fullscreen background color 上游逐字);申辩:透明图全屏 matte 不重合成(入册偏离,上游自家 TODO viv.c:224 同病)、页面重组/组合只列已实现值(issue 明文)、paint FitPolicy 重复+diag 残留+ViewOptions 灰化(已在 amend 0e8d052 修复,cubic 审的是旧 head)。216 绿 + smoke24 复跑 48 PASS

+++ src/fit.rs
@@ -49,10 +49,21 @@ pub(crate) fn window_size_client(
         1 => image,
         2 => (image.0.saturating_mul(2), image.1.saturating_mul(2)),
         // A zero divisor makes upstream skip that axis (rect stays 0);
-        // the clamp then keeps whatever the other axis produced.
+        // the clamp then keeps whatever the other axis produced. The
+        // multiply runs in i64 — a hand-edited huge `auto_fit_*_mul`
+        // overflows i32 before the division (upstream wraps as C;
+        // riviv keeps the widened value honest instead).
         _ => (
-            if wd != 0 { full_monitor.0 * wm / wd } else { 0 },
-            if hd != 0 { full_monitor.1 * hm / hd } else { 0 },
+            if wd != 0 {
+                (i64::from(full_monitor.0) * i64::from(wm) / i64::from(wd)) as i32
+            } else {
+                0
+            },
+            if hd != 0 {
+                (i64::from(full_monitor.1) * i64::from(hm) / i64::from(hd)) as i32
+            } else {
+                0
+            },
         ),
     };
     (w.clamp(0, work.0), h.clamp(0, work.1))
